% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_matching.R
\name{accuracy}
\alias{accuracy}
\title{Calculate Mean Accuracy and Mean Squared Accuracy of ID Matching}
\usage{
accuracy(ids)
}
\arguments{
\item{ids}{A data frame with matched IDs.}
}
\value{
A message displaying the calculated mean accuracy and mean squared
accuracy. The result is printed to the console.
}
\description{
This function takes ID matching results produced by \code{\link[=match_ids]{match_ids()}} and
calculates the mean accuracy and mean squared accuracy of the matches.
}
