% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoo-centrix.R
\name{aspect}
\alias{aspect}
\title{Create new aspect object}
\usage{
aspect(x = integer(), levels = c("G", "YY", "Y", "R"))
}
\arguments{
\item{x}{Aspect}

\item{levels}{Aspect levels}
}
\description{
Create new aspect object
}
