% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-data.R
\docType{data}
\name{aspect_events}
\alias{aspect_events}
\title{Example aspect events}
\format{
\describe{
\item{signal}{Signal ID}
\item{dt}{Date and time of event}
\item{aspect}{Signal aspect caused by event}
\item{past_aspect}{Signal aspect preceding event}
}
}
\usage{
aspect_events
}
\description{
An example of Centrix data which have been pre-processed into aspect events.
}
\examples{
aspect_events

}
\seealso{
\code{\link[=preprocess_signal_events]{preprocess_signal_events()}}

A data frame containing 32,512 rows and 4 columns.
}
\keyword{datasets}
