% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asset_map}
\alias{asset_map}
\title{Example asset map}
\format{
\describe{
\item{signal}{Signal ID.}
\item{berth}{Berth ID.}
\item{track}{Track ID.}
\item{event}{Track event.}
}
}
\usage{
asset_map
}
\description{
An example of an asset map showing the relationship between different track
assets. There is a 1-1 mapping from signal to berth, a 1-many mapping from
berth to track, and a 1-1 mapping from track to event.
}
\details{
A data frame containing 17 rows and 4 columns.
}
\examples{
asset_map

}
\keyword{datasets}
