% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-berth.R
\name{berth}
\alias{berth}
\title{Generate a new berth object}
\usage{
berth(
  id = character(),
  signal = character(),
  state = aspect(),
  station = character(),
  length = integer(),
  L1 = integer(),
  L2 = integer()
)
}
\arguments{
\item{id}{Berth name}

\item{signal}{Signal name}

\item{state}{Signal aspect}

\item{station}{Station name}

\item{length}{Berth length in metres}

\item{L1}{Distance from start of berth to station in metres}

\item{L2}{Distance from station to end of berth in metres}
}
\description{
Generate a new berth object
}
