% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-data.R
\docType{data}
\name{berth_events}
\alias{berth_events}
\title{Example of berth-level events}
\format{
\describe{
\item{signal}{Signal ID}
\item{berth}{Berth ID}
\item{train_id}{Train ID}
\item{aspect}{Signal aspect as train enters berth}
\item{t_enters}{Date and time that train enters berth}
\item{t_red_on}{Date and time that signal aspect changes to red}
\item{t_enters_next}{Date and time that train enters next berth}
\item{t_vacates}{Date and time that train vacates berth}
\item{t_red_off}{Date and time that signal aspect changes from red}
\item{TSAR}{Amount of time signal aspect is set to red}
\item{T_onset}{Time between train entering berth and signal aspect changing
to red}
\item{T_clear}{Amount of time taken for train to fully clear the berth}
\item{T_offset}{Time between train vacating berth and signal aspect
changing from red}
\item{T_travel}{Amount of time taken for train to travel the length of the
berth}
\item{T_coach}{Amount of time taken for train to travel its own length}
}
}
\usage{
berth_events
}
\description{
An example of processed Centrix data containing TSAR and its sub-components.
}
\examples{
berth_events

}
\seealso{
\code{\link[=calculate_tsars]{calculate_tsars()}} \code{\link[=wrangle_centrix]{wrangle_centrix()}} \code{\link[=cluster_journeys]{cluster_journeys()}}

A data frame containing 6,952 rows and 15 columns
}
\keyword{datasets}
