% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-data.R
\docType{data}
\name{berth_lengths}
\alias{berth_lengths}
\title{Example berth lengths}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 6 columns.
}
\usage{
berth_lengths
}
\description{
Example berth lengths
}
\keyword{datasets}
