% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix.R
\name{calculate_time_windows}
\alias{calculate_time_windows}
\title{Calculate Centrix Time Windows}
\usage{
calculate_time_windows(aspect_events, track_events, asset_map)
}
\arguments{
\item{aspect_events}{A data frame containing pre-processed aspect events from
Centrix data. No input validation is performed. It is assumed that the data
frame will follow the structure returned by the function
\code{\link[=preprocess_signal_events]{preprocess_signal_events()}}.}

\item{track_events}{A data frame containing pre-processed track events from
Centrix data. No input validation is performed. It is assumed that the data
frame will follow the structure return by the function
\code{\link[=preprocess_track_events]{preprocess_track_events()}}.}

\item{asset_map}{A data frame containing a map of the track section. No input
validation is performed. See \code{\link[=wrangle_centrix]{wrangle_centrix()}} for the expected
structure.}
}
\value{
A data frame containing window IDs and time intervals in the
columns:
\itemize{
\item window: (numeric) window ID,
\item interval: (lubridate::interval) window time interval.
Each window indicates a time interval in which the data contain valid
observations.
}
}
\description{
Calculates a set of time windows containing valid Centrix observations. This
function acts as a wrapper for a number of functions involved in the
calculation. Returns a data frame containing the window ID and time interval.
}
