% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-clustering.R
\name{cluster_centrix}
\alias{cluster_centrix}
\title{Cluster Centrix Data}
\usage{
cluster_centrix(
  berth_events,
  outliers = NULL,
  outlier_detection = "none",
  k = 3L,
  niter = 20L
)
}
\arguments{
\item{berth_events}{A data frame containing berth-level Centrix events
containing the columns: signal, berth, train_id, aspect, T_travel,
time_elapsed.}

\item{outliers}{A data frame containing the train IDs for any outliers. Only
used if \code{outlier_detection} == "manual".}

\item{outlier_detection}{Which approach to use for outliers. Options are:
\itemize{
\item{none} No outliers removed.
\item{manual} Manually remove outliers using the \code{outliers} parameter.
\item{boxplot} Unimplemented.
}}

\item{k}{How many clusters to find.}

\item{niter}{How many iterations to use for the clustering.}
}
\value{
A data frame containing train IDs and their clusters. Clusters are
ordered based on total variance, highest variance first.
}
\description{
Clusters Centrix data into groups based on travel times across berths.
}
