% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-clustering.R
\name{cluster_journeys}
\alias{cluster_journeys}
\title{Cluster train journeys}
\usage{
cluster_journeys(
  berth_events,
  outliers = NULL,
  outlier_detection = "none",
  ...
)
}
\arguments{
\item{berth_events}{A data frame containing berth-level Centrix events
containing the columns: signal, berth, train_id, aspect, T_travel,
time_elapsed.}

\item{outliers}{A data frame containing the train IDs for any outliers.
Ignored unless \code{outlier_detection} == "manual".}

\item{outlier_detection}{Which approach to use for outliers. Options are:
\itemize{
\item{"none"} No outliers removed.
\item{"manual"} Manually remove outliers using the \code{outliers} parameter.
\item{"boxplot"} Unimplemented.
}}

\item{...}{
  Arguments passed on to \code{\link[stats:kmeans]{stats::kmeans}}
  \describe{
    \item{\code{x}}{numeric matrix of data, or an object that can be coerced to
    such a matrix (such as a numeric vector or a data frame with all
    numeric columns).}
    \item{\code{centers}}{either the number of clusters, say \eqn{k}, or a set of
    initial (distinct) cluster centres.  If a number, a random set of
    (distinct) rows in \code{x} is chosen as the initial centres.}
    \item{\code{iter.max}}{the maximum number of iterations allowed.}
    \item{\code{nstart}}{if \code{centers} is a number, how many random sets
    should be chosen?}
    \item{\code{algorithm}}{character: may be abbreviated.  Note that
    \code{"Lloyd"} and \code{"Forgy"} are alternative names for one
    algorithm.}
    \item{\code{trace}}{logical or integer number, currently only used in the
    default method (\code{"Hartigan-Wong"}): if positive (or true),
    tracing information on the progress of the algorithm is
    produced.  Higher values may produce more tracing information.}
  }}
}
\value{
A data frame containing the columns:
\itemize{
\item{\code{signal}}{Signal ID}
\item{\code{berth}}{Berth ID}
\item{\code{train_id}}{Train ID}
\item{\code{aspect}}{Signal aspect as train enters berth}
\item{\code{T_travel}}{Amount of time taken to travel the length of the berth}
\item{\code{cluster}}{Cluster ID}
}
}
\description{
Clusters Centrix data into groups based on travel times across berths. This
function makes use of K-means clustering, which is inherently
non-deterministic. Therefore, the output of this function is not always what
would be expected. It is recommended to check the output visually using
\code{\link[=plot_clusters]{plot_clusters()}}. For reproducability, clusters are ordered based on total
variance, highest variance first.
}
\examples{
data(berth_events)
berth_events_clusters <- cluster_journeys(berth_events,
                                          centers = 3L,
                                          iter.max = 100L)
plot_clusters(berth_events_clusters)

}
\seealso{
\code{\link[stats:kmeans]{stats::kmeans()}} \code{\link[=wrangle_centrix]{wrangle_centrix()}} \code{\link[=plot_clusters]{plot_clusters()}}
\code{\link[=match_ids]{match_ids()}} \code{\link[=estimate_berth_lengths]{estimate_berth_lengths()}} \code{\link[=estimate_dwell_times]{estimate_dwell_times()}}
\link{berth_events} \link{berth_events_groups}
}
