% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_matching.R
\name{combine_data}
\alias{combine_data}
\title{Combine Observed and Timetable Data based on ID Matching}
\usage{
combine_data(ids, observed, timetable)
}
\arguments{
\item{ids}{A data frame containing the results of ID matching. It should have
the following columns:
\itemize{
\item \code{train_id}: Integer, the train ID obtained from ID matching.
\item \code{train_header}: Character, the corresponding train header
from the timetable data.
}}

\item{observed}{A data frame containing observed Centrix data that has been
pre-processed for ID matching. It should contain the following columns:
\itemize{
\item \code{train_id}: Integer vector, train ID.
\item \code{signal}: Character vector, signal ID.
\item \code{t_enters}: \link{POSIXct}, the time when the train enters the
berth.
\item \code{t_vacates}: \link{POSIXct}, the time when the train vacates the
berth.
\item \code{group}: Character vector, the group name.
}}

\item{timetable}{A data frame containing pre-processed timetable data with
the following columns:
\itemize{
\item \code{train_header}: Character vector, the train header.
\item \code{geo}: Character vector, the geographical location.
\item \code{t}: \link{POSIXct}, the actual time of the event.
\item \code{event}: Character vector, the type of event.
\item \code{group}: Character vector, the group name.
}}
}
\value{
A data frame containing the combined information from the ID matching
results, observed data, and timetable data. The data frame is sorted by
train IDs in ascending order.
}
\description{
This function combines observed and timetable data based on the results of ID
matching. It takes three data frames as input: \code{ids}, \code{observed}, and
\code{timetable}, and returns a combined data frame with relevant information.
}
