% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-data.R
\docType{data}
\name{dwell_times}
\alias{dwell_times}
\title{Example dwell times}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1581 rows and 21 columns.
}
\usage{
dwell_times
}
\description{
Example dwell times
}
\keyword{datasets}
