% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-berth_lengths.R
\name{estimate_berth_lengths}
\alias{estimate_berth_lengths}
\title{Estimate berth lengths}
\usage{
estimate_berth_lengths(
  timetable_specification,
  id_matching,
  berth_events_groups,
  expected_journey_time,
  track_length
)
}
\arguments{
\item{timetable_specification}{Timetable specification}

\item{id_matching}{Id matching}

\item{berth_events_groups}{Berth events groups}

\item{expected_journey_time}{Expected journey time in seconds from start of
Centrix track to end}

\item{track_length}{Total length of Centrix track in miles}
}
\description{
Estimate berth lengths
}
