% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-delays.R
\name{estimate_delays}
\alias{estimate_delays}
\title{Calculate delays from Centrix and timetable data}
\usage{
estimate_delays(dwell_times, timetable_subset, id_matching)
}
\arguments{
\item{dwell_times}{Data frame containing estimated dwell times. The structure
is expected to match the output of \code{\link[=estimate_dwell_times]{estimate_dwell_times()}}. However, only
6 columns are required for this function to work:
\itemize{
\item{\code{signal}}{Signal ID}
\item{\code{train_id}}{Train ID (Centrix)}
\item{\code{station}}{Station name}
\item{\code{t_enters}}{Date and time that train enters berth}
\item{\code{T_move_one}}{Moving time before coming to a stop at the station}
\item{\code{T_dwell}}{Dwell time at the station}
}}

\item{timetable_subset}{Processed timetable containing calling patterns. The
structure is expected to match the output of \code{\link[=wrangle_timetable]{wrangle_timetable()}}.
However, only 7 columns are required:
\itemize{
\item{\code{train_header}}{Train ID (timetable)}
\item{\code{dt_origin}}{Date and time of train origin}
\item{\code{geo}}{Name of location}
\item{\code{group}}{Group name}
\item{\code{event}}{Event name; anything other than 'Arrive' or 'Depart' is
ignored}
\item{\code{wtt}}{Scheduled time of event}
\item{\code{t}}{Actual time of event}
}}

\item{id_matching}{Data frame containing a 1-1 mapping from Centrix IDs to
timetable IDs. There should be 3 columns:
\itemize{
\item{\code{train_id}}{Train ID (Centrix)}
\item{\code{train_header}}{Train ID (timetable)}
\item{\code{dt_origin}}{Date and time of train origin}
}}
}
\value{
A data frame containing 8 columns:
\itemize{
\item{\code{signal}}{Signal ID}
\item{\code{train_id}}{Train ID (Centrix)}
\item{\code{station}}{Station name}
\item{\code{T_dwell}}{Dwell time in seconds}
\item{\code{arrival_delay}}{Arrival delay in seconds, can be negative if the
train arrives early}
\item{\code{departure_delay}}{Departure delay in seconds, can be negative if
the train departs early}
\item{\code{excess_dwell_time}}{Difference between actual dwell time and
scheduled dwell time. Can be negative if the
train spends less time in the station than
scheduled}
\item{\code{early}}{Logical vector indicating whether the train arrived at the
station earlier than scheduled}
}
}
\description{
Calculate delays from estimated dwell times and timetable data. This function
calculates arrival delay, departure delay, excess dwell time, and whether a
train was early or late. These are all estimates, as the dwell times on which
they are based are also estimates.
}
\examples{
data(dwell_times, timetable_subset, id_matching)
dwell_times
timetable_subset
id_matching
delays <- estimate_delays(dwell_times, timetable_subset, id_matching)
delays

}
\seealso{
\code{\link[=estimate_dwell_times]{estimate_dwell_times()}} \code{\link[=wrangle_timetable]{wrangle_timetable()}} \code{\link[=match_ids]{match_ids()}}
\link{dwell_times}
}
