% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{event_mapping}
\alias{event_mapping}
\title{Default Event Mapping}
\format{
\subsection{\code{event_mapping}}{

A data frame with 5 rows and 2 columns:
\describe{
\item{event}{Raw event code}
\item{name}{Event name}
}
}
}
\usage{
event_mapping
}
\description{
Default Event Mapping
}
\seealso{
\code{\link[=wrangle_timetable]{wrangle_timetable()}}
}
\keyword{datasets}
