% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{filter_relevant_direction}
\alias{filter_relevant_direction}
\title{Filter timetable by direction}
\usage{
filter_relevant_direction(timetable, stations)
}
\arguments{
\item{timetable}{Data frame containing timetable data.}

\item{stations}{List containing station names.}
}
\value{
A (usually) smaller timetable only containing trains which pass
through or stop at the specified stations in the given order.
}
\description{
Filter timetable by direction
}
