% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-windows.R
\name{filter_track_events}
\alias{filter_track_events}
\title{Filter Track Events}
\usage{
filter_track_events(track_events, time_windows, asset_map)
}
\arguments{
\item{track_events}{A data frame containing pre-processed track events from
Centrix data. No input validation is performed. It is assumed that the data
frame will follow the structure return by the function
\code{\link[=preprocess_track_events]{preprocess_track_events()}}.}

\item{time_windows}{A data frame containing window IDs and time intervals in
the columns:
\itemize{
\item window: (numeric) window ID,
\item interval: (lubridate::interval) window time interval.
No input validation is performed. The structure is expected to follow the
output of \code{\link[=calculate_time_windows]{calculate_time_windows()}}
}}

\item{asset_map}{A data frame containing a map of the track section. No input
validation is performed. See \code{\link[=wrangle_centrix]{wrangle_centrix()}} for the expected
structure.}
}
\value{
A data frame containing filtered track events, with additional
columns:
\itemize{
\item window: (numeric) the window ID as calculated in \code{\link[=calculate_time_windows]{calculate_time_windows()}}
\item interval: (lubridate::interval) the time interval of the window
}
}
\description{
Filters Centrix track events to only the events that lie within the
calculated valid time windows. Also filters to only the tracks specified in
the asset map.
}
