% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{find_calling_patterns}
\alias{find_calling_patterns}
\title{Find calling patterns from timetable}
\usage{
find_calling_patterns(timetable)
}
\arguments{
\item{timetable}{Data frame containing timetable data.}
}
\value{
The timetable with calling patterns added.
}
\description{
Find calling patterns from timetable
}
