% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_asset_mapping}
\alias{get_asset_mapping}
\title{Get the Asset Mapping}
\usage{
get_asset_mapping()
}
\value{
A list containing the asset mapping.
}
\description{
This function retrieves the asset mapping from the current environment. The
asset mapping is stored in the environment variable 'asset_mapping'.
}
\examples{
mapping <- get_asset_mapping()
print(mapping)

}
\seealso{
Use \link{set_asset_mapping} to set the asset mapping in the environment.
}
