% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_centrix}
\alias{get_centrix}
\title{Get Centrix Data from Environment}
\usage{
get_centrix()
}
\value{
A data frame containing the centrix data.
}
\description{
This function retrieves the centrix data from the current environment. The
centrix data is stored in the environment variable 'centrix'.
}
\examples{
centrix <- get_centrix()
print(centrix)

}
\seealso{
Use \link{set_centrix} to set centrix data in the environment.
}
