% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_timetable.R
\name{get_event_mapping}
\alias{get_event_mapping}
\title{Get Event Mapping}
\usage{
get_event_mapping()
}
\value{
The event mapping table as a tibble data frame.
}
\description{
This function retrieves the current event mapping table.
}
\details{
The event mapping table is stored in the environment object named "env".
}
\seealso{
\link{set_event_mapping}
}
