% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aac-internals.R
\name{get_map}
\alias{get_map}
\title{Retrieve network map}
\usage{
get_map()
}
\description{
\code{get_map()} retrieves the network map from the internal environment. By
default, this will equal \code{NULL}. Use \code{\link[=set_map]{set_map()}} to store a map.
}
