% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_state_mapping}
\alias{get_state_mapping}
\title{Get State Mapping from Environment}
\usage{
get_state_mapping()
}
\value{
A data frame containing the state mapping.
}
\description{
This function retrieves the signal state mapping from the current
environment. The mapping is stored in the environment variable
'state_mapping'.
}
\examples{
state_mapping <- get_state_mapping()
print(state_mapping)

}
\seealso{
Use \link{set_state_mapping} to set state mapping in the environment.
}
