% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_stations}
\alias{get_stations}
\title{Get Stations from Environment}
\usage{
get_stations()
}
\description{
This function retrieves the currently stored list of stations names from the
environment.
}
\seealso{
\code{\link[=set_stations]{set_stations()}}
}
