% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_timetable}
\alias{get_timetable}
\title{Get Timetable Data from Environment}
\usage{
get_timetable()
}
\value{
A data frame containing the timetable data.
}
\description{
This function retrieves the timetable data from the current environment. The
timetable data is stored in the environment variable 'timetable'.
}
\examples{
timetable <- get_timetable()
print(timetable)

}
\seealso{
Use \link{set_timetable} to set timetable data in the environment.
}
