% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching-data.R
\docType{data}
\name{id_matching}
\alias{id_matching}
\title{Example ID Matching}
\format{
\describe{
\item{train_id}{Centrix train ID}
\item{train_header}{Timetable ID}
\item{dt_origin}{Datetime that train originated}
}
}
\usage{
id_matching
}
\description{
Data frame containing matched Centrix and timetable IDs
}
\details{
Data frame with 42 rows and 3 columns
}
\keyword{datasets}
