% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_matching.R
\name{match_ids}
\alias{match_ids}
\title{Match Centrix and Timetable IDs}
\usage{
match_ids(berth_events_classes, timetable_groups, match_mapping)
}
\arguments{
\item{berth_events_classes}{Data frame containing berth-level Centrix data.}

\item{timetable_groups}{Data frame containing Timetable data.}

\item{match_mapping}{Data frame containing instructions for the matching.}
}
\value{
Data frame containing a 1-1 mapping between Centrix IDs and
Timetable IDs.
}
\description{
Calculates matching IDs between Centrix and Timetable data.
}
