% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_matching.R
\name{preprocess_berths}
\alias{preprocess_berths}
\title{Preprocess Berth Data for ID Matching}
\usage{
preprocess_berths(berth_events, train_classes)
}
\arguments{
\item{berth_events}{Data frame containing berth data}

\item{train_classes}{Data frame containing train classes}
}
\description{
Preprocess Berth Data for ID Matching
}
