% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_centrix.R
\name{preprocess_signal_events}
\alias{preprocess_signal_events}
\title{Preprocess raw signal events}
\usage{
preprocess_signal_events(raw_signal_events, state_mapping)
}
\description{
Preprocess raw signal events
}
