% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_timetable.R
\name{preprocess_timetable}
\alias{preprocess_timetable}
\title{Reads and pre-processes timetable data}
\usage{
preprocess_timetable(path, names, ...)
}
\arguments{
\item{path}{String pointing to the directory containing raw timetable data}

\item{names}{Named list containing column names and what to rename them to}

\item{...}{Additional arguments passed to \code{\link{read_excel_files}}}
}
\description{
Reads excel data from a given path and renames the columns according to
\code{names}
}
