% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/railvarr-package.R
\docType{package}
\name{railvarr-package}
\alias{railvarr}
\alias{railvarr-package}
\title{railvarr: Rail Network Performance Variability Analysis}
\description{
Provide functionality to aid in the pre-processing and analysis of railway data, with the goal of analysing network performance variability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebmerricks/railvarr}
  \item \url{https://sebmerricks.github.io/railvarr/}
  \item Report bugs at \url{https://github.com/sebmerricks/railvarr/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Merricks \email{b8072600@newcastle.ac.uk}

}
\keyword{internal}
