% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_centrix.R
\name{read_centrix}
\alias{read_centrix}
\title{Read Centrix Data From Multiple Files}
\usage{
read_centrix(path, names, types)
}
\arguments{
\item{path}{A string pointing to the directory pointing to the relevant files.}

\item{names}{A character vector containing the column names.}

\item{types}{A column specification created with \code{\link[readr]{cols}}.}
}
\description{
Read Centrix Data From Multiple Files
}
