% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_excel_files}
\alias{read_excel_files}
\title{Read raw excel data from multiple files}
\usage{
read_excel_files(path, progress = TRUE, ...)
}
\arguments{
\item{path}{A character string specifying the path to the directory
containing the excel files.}

\item{progress}{Whether to show a progress bar, see \code{\link[purrr:map]{purrr::map()}}.}

\item{...}{Additional arguments to be passed to \code{\link[readxl:read_excel]{readxl::read_excel()}}.}
}
\value{
A single data frame.
}
\description{
Reads all the excel files contained in the directory specified by \code{path} and
combines them into a single data frame. Errors will be thrown if either
the directory does not exist or is empty.
}
\seealso{
\code{\link[readxl:read_excel]{readxl::read_excel()}} which this function wraps.
}
