% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{set_centrix}
\alias{set_centrix}
\title{Set Centrix in Environment with Validation}
\usage{
set_centrix(new_centrix)
}
\arguments{
\item{new_centrix}{A data frame representing the Centrix data containing
columns 'asset', 'dt', 'transition', and 'period'. The input Centrix must
adhere to the expected format and validation rules.}
}
\description{
This function allows you to set the Centrix data in the current environment
after validating the format of the provided data. The data are stored in the
environment variable 'centrix'.
}
\details{
This function performs strict validation on the input \code{new_centrix}
using predefined rules. The function checks if the provided data adhere
to the expected format and throws an error if any of the validation rules
are violated.

The input \code{new_centrix} must be a data frame with the following elements:
\itemize{
\item \code{asset}: A character vector representing the asset ID in the
format "(S|T).+" (e.g., "S123" or "TABC").
\item \code{dt}: A POSIXct object representing the date and time of the
observation.
\item \code{transition}: A character vector representing the transition
type, which must be either "DN to UP" or "UP to DN".
\item \code{period}: A numeric value representing the time period of the
Centrix.
}
}
