% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_timetable.R
\name{set_event_mapping}
\alias{set_event_mapping}
\title{Set Event Mapping}
\usage{
set_event_mapping(mapping)
}
\arguments{
\item{mapping}{A tibble data frame representing the event mapping.
The data frame should have columns "event" and "name".
The default value is a predefined event mapping.}
}
\description{
This function sets the event mapping table in the environment.
}
\examples{
# Set a custom event mapping
custom_mapping <- dplyr::tribble(
  ~event, ~name,
  "A", "Arrive",
  "D", "Depart"
)
set_event_mapping(custom_mapping)

}
