% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_centrix.R
\name{set_map}
\alias{set_map}
\title{Set network map}
\usage{
set_map(map)
}
\arguments{
\item{map}{New network map to set. Should be in order: the first entry is the
start of the track, while the last entry is the end.}
}
\description{
Set network map
}
