% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{set_state_mapping}
\alias{set_state_mapping}
\title{Set Signal State Mapping in Environment}
\usage{
set_state_mapping(new_state_mapping)
}
\arguments{
\item{new_state_mapping}{A data frame representing the state mapping
containing columns 'signal' and 'aspect'. The input mapping must adhere to
the expected format and validation rules.}
}
\description{
This function allows you to set the state mapping in the current environment
after validating the format of the provided mapping. The mapping is stored in
the environment variable 'state_mapping'.
}
\details{
This function performs strict validation on the input
\code{new_state_mapping} using predefined rules. The function checks if the
provided data adhere to the expected format and throws an error if any of
the validation rules are violated.

The input \code{new_state_mapping} must be a data frame with the following
elements:
\itemize{
\item \code{state}: A character vector representing the signal state as
recorded in the raw Centrix data. e.g., 'RGE' or 'DGE'.
\item \code{aspect}: A character vector representing the signal aspect,
e.g., 'R' or 'G'.
}
This mapping should be a 1-1 mapping from state to aspect.
}
