% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{set_stations}
\alias{set_stations}
\title{Store Station Names in Environment}
\usage{
set_stations(new_stations)
}
\arguments{
\item{new_stations}{A list of character vectors containing the station names.}
}
\description{
This function sets the station names in the environment for use in timetable
filtering.
}
\seealso{
\code{\link[=get_stations]{get_stations()}}
}
