% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{set_timetable}
\alias{set_timetable}
\title{Set Timetable in Environment with Validation}
\usage{
set_timetable(new_timetable)
}
\arguments{
\item{new_Timetable}{A data frame representing the timetable data containing
columns 'train_header', 'dt_origin', 'geo', 'event', 'wtt', 't', and
'delay'. The input Timetable must adhere to the expected format and
validation rules.}
}
\description{
This function allows you to set the timetable data in the current environment
after validating the format of the provided data. The data are stored in the
environment variable 'timetable'.
}
\details{
This function performs strict validation on the input
\code{new_timetable} using predefined rules. The function checks if the provided
data adhere to the expected format and throws an error if any of the
validation rules are violated.

The input \code{new_timetable} must be a data frame with the following elements:
\itemize{
\item \code{train_header}: A character vector representing the train ID.
\item \code{dt_origin}: A POSIXct object representing the date and time at
which the train departed the origin.
\item \code{geo}: A character vector representing the location of the
timetable event.
\item \code{event}: A character vector representing the timetabling event.
See \link{get_event_mapping} for what these events can be.
\item \code{wtt}: A POSIXct object representing the scheduled time.
\item \code{t}: A POSIXct object representing the actual time.
\item \code{delay}: A numeric value representing the delay.
}
}
