% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoo-centrix.R
\name{signal_event}
\alias{signal_event}
\title{Create new signal event object}
\usage{
signal_event(
  signal_id = character(),
  datetime = lubridate::POSIXct(),
  previous_aspect = aspect(),
  new_aspect = aspect()
)
}
\arguments{
\item{signal_id}{Signal ID}

\item{datetime}{Datetime}

\item{previous_aspect}{Previous aspect}

\item{new_aspect}{New aspect}
}
\description{
Create new signal event object
}
