% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_centrix.R
\name{split_signal_track_events}
\alias{split_signal_track_events}
\title{Split raw events into separate signal and track events}
\usage{
split_signal_track_events(raw_events)
}
\arguments{
\item{raw_events}{Data frame containing raw data.}

\item{is_track}{Expression determining how to split tracks and signals.}
}
\description{
Split raw events into separate signal and track events
}
