% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state_mapping}
\alias{state_mapping}
\title{Default signal state to aspect mapping}
\format{
\subsection{\code{state_mapping}}{

A data frame with 4 rows and 2 columns:
\describe{
\item{state}{Signal state}
\item{aspect}{Signal aspect as factor}
}
}
}
\usage{
state_mapping
}
\description{
The default 1-1 mapping from signal state to aspect. See
\url{https://wiki.openraildata.com/index.php/Signalling_Nomenclature#Signals}.
This is used as the default state mapping in Centrix wrangling.
}
\seealso{
\code{\link[=wrangle_centrix]{wrangle_centrix()}}
}
\keyword{datasets}
