% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable-data.R
\docType{data}
\name{stopping_stations}
\alias{stopping_stations}
\title{Example stopping stations}
\format{
An object of class \code{list} of length 3.
}
\usage{
stopping_stations
}
\description{
List containing stations at which trains stop. This is used for finding
calling patterns
}
\details{
A list with 3 elements.
}
\examples{
stopping_stations

}
\seealso{
\code{\link[=wrangle_timetable]{wrangle_timetable()}} \code{\link[=find_calling_patterns]{find_calling_patterns()}}
}
\keyword{datasets}
