% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-data.R
\docType{data}
\name{time_windows}
\alias{time_windows}
\title{Example time windows}
\format{
\describe{
\item{window}{Window ID}
\item{interval}{Window time interval}
}
}
\usage{
time_windows
}
\description{
An example of time windows calculated from Centrix data denoting time
intervals of valid train journeys.
}
\examples{
time_windows

}
\seealso{
\code{\link[=calculate_time_windows]{calculate_time_windows()}}

A data frame containing 681 rows and 2 columns
}
\keyword{datasets}
