% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable-data.R
\docType{data}
\name{timetable}
\alias{timetable}
\title{Example timetable data}
\format{
\describe{
\item{train_header}{Train ID. This usually contains headcode information}
\item{dt_origin}{Date and time of service origin}
\item{geo}{Name of TIPLOCs that trains pass through}
\item{event}{Type of timetable event, e.g., 'Arrive', 'Depart', 'Pass'}
\item{wtt}{Scheduled date and time of event}
\item{t}{Actual date and time of event}
\item{delay}{Difference between actual time and scheduled time}
\item{allow}{Total delay allowance}
\item{allow_perf}{Performance allowance}
\item{allow_path}{Pathing allowance}
\item{allow_eng}{Engineering allowance}
}
}
\usage{
timetable
}
\description{
Raw timetable data example. This is an anonymised version of a real data set.
}
\details{
A data frame with 8,958 rows and 11 columns
}
\examples{
timetable

}
\seealso{
\code{\link[=wrangle_timetable]{wrangle_timetable()}}
}
\keyword{datasets}
