% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching-data.R
\docType{data}
\name{timetable_groups}
\alias{timetable_groups}
\title{Example of a wrangled timetable}
\format{
\describe{
\item{train_header}{Train ID. This usually contains headcode information}
\item{dt_origin}{Date and time of service origin}
\item{geo}{Name of TIPLOCs that trains pass through or stop at}
\item{event}{Type of timetable event, e.g., 'Arrive', 'Depart', 'Pass'}
\item{wtt}{Scheduled date and time of event}
\item{t}{Actual date and time of event}
\item{delay}{Difference between actual time and scheduled time}
\item{allow}{Total delay allowance}
\item{allow_perf}{Performance allowance}
\item{allow_path}{Pathing allowance}
\item{allow_eng}{Engineering allowance}
\item{group}{Group name corresponding to calling pattern}
}
}
\usage{
timetable_groups
}
\description{
Timetable containing only relevant services travelling in a single direction
with calling patterns labelled.
}
\details{
A data frame with 792 rows and 12 columns
}
\keyword{datasets}
