% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable-data.R
\docType{data}
\name{timetable_specification}
\alias{timetable_specification}
\title{Example timetable specification}
\format{
\describe{
\item{train_header}{Train ID. This usually contains headcode information}
\item{dt_origin}{Date and time of service origin}
\item{wtt}{Scheduled date and time of event}
\item{group}{Group name}
\item{xi}{Start point}
\item{xk}{End point}
\item{duration}{Scheduled time between \code{xi} and \code{xk}}
\item{type}{Type of travel, \code{moving} or \code{dwell}}
\item{j}{Segment ID}
\item{T_journey}{Total journey time across track section}
}
}
\usage{
timetable_specification
}
\description{
Timetable specification containing information about expected journey times
}
\details{
Data frame with 281 rows and 10 columns
}
\examples{
timetable_specification

}
\seealso{
\code{\link[=calculate_journey_specifications]{calculate_journey_specifications()}} \code{\link[=match_ids]{match_ids()}}
\code{\link[=estimate_berth_lengths]{estimate_berth_lengths()}}
}
\keyword{datasets}
