% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable-data.R
\docType{data}
\name{timetable_subset}
\alias{timetable_subset}
\title{Example timetable subset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 792 rows and 12 columns.
}
\usage{
timetable_subset
}
\description{
Example timetable subset
}
\keyword{datasets}
