% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoo-centrix.R
\name{track_event}
\alias{track_event}
\title{Create new track event}
\usage{
track_event(
  track_id = character(),
  datetime = lubridate::POSIXct(),
  event_type = character()
)
}
\arguments{
\item{track_id}{Track ID}

\item{datetime}{Datetime}

\item{event_type}{Event type ('enters'/'vacates')}
}
\description{
Create new track event
}
