% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-tsar_event.R
\name{tsar_event}
\alias{tsar_event}
\title{Create a new tsar event}
\usage{
tsar_event(
  t_enters = lubridate::POSIXct(),
  t_red_on = lubridate::POSIXct(),
  t_enters_next = lubridate::POSIXct(),
  t_vacates = lubridate::POSIXct(),
  t_red_off = lubridate::POSIXct()
)
}
\arguments{
\item{t_enters}{Time train enters berth}

\item{t_red_on}{Time signal changes to red}

\item{t_enters_next}{Time train enters next berth}

\item{t_vacates}{Time train vacates berth}

\item{t_red_off}{Time signal changes off red}
}
\description{
Create a new tsar event
}
