% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_aspect_events}
\alias{valid_aspect_events}
\title{Example of processed aspect events}
\format{
\describe{
\item{signal}{Signal ID}
\item{dt}{Date and time of event}
\item{aspect}{Signal aspect caused by event}
\item{past_aspect}{Signal aspect preceding event}
\item{window}{Window ID}
\item{interval}{Window interval}
}
}
\usage{
valid_aspect_events
}
\description{
An example of aspect events which have been fully processed and validated by
the calculated time windows.
}
\details{
A data frame containing 25,834 rows and 6 columns
}
\keyword{datasets}
