% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-data.R
\docType{data}
\name{valid_track_events}
\alias{valid_track_events}
\title{Example of processed track events}
\format{
\describe{
\item{track}{Track ID}
\item{dt}{Date and time of event}
\item{occupied}{TRUE if train enters track, else FALSE}
\item{event}{'enters' if train enters track, else 'vacates'}
\item{window}{Window ID}
\item{interval}{Window interval}
}
}
\usage{
valid_track_events
}
\description{
An example of track events which have been fully processed and validated by
the calculated time windows.
}
\details{
A data frame containing 24,332 rows and 6 columns
}
\keyword{datasets}
