% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_centrix.R
\name{wrangle_centrix}
\alias{wrangle_centrix}
\title{Wrangle raw Centrix data}
\usage{
wrangle_centrix(raw_signal_events, raw_track_events)
}
\arguments{
\item{raw_signal_events}{Data frame containing raw signal data.}

\item{raw_track_events}{Data frame containing raw track data.}
}
\description{
Wrangle raw Centrix data
}
