% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-aspect.R
\name{aspect}
\alias{aspect}
\title{Create new aspect}
\usage{
aspect(x = character(), levels = c("R", "Y", "YY", "G"))
}
\arguments{
\item{x}{Aspect}

\item{levels}{Possible aspects}
}
\description{
Create new aspect
}
