% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-berth.R
\name{berth}
\alias{berth}
\title{Generate a new berth object}
\usage{
berth(
  id = character(),
  signal = character(),
  state = aspect(),
  station = character(),
  length = numeric()
)
}
\arguments{
\item{id}{Berth name}

\item{signal}{Signal name}

\item{state}{Signal aspect}

\item{station}{Station name}

\item{length}{Berth length}
}
\description{
Generate a new berth object
}
