% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-data.R
\docType{data}
\name{berth_lengths}
\alias{berth_lengths}
\title{Example berth lengths}
\format{
\describe{
\item{berth}{Berth ID}
\item{median_travel}{Median travel time for each berth}
\item{v.mph}{Expected speed across berth}
\item{L.miles}{Estimated berth length in miles}
\item{L.km}{Estimated berth length in kilometres}
\item{L}{Estimated berth length in metres}
}
}
\usage{
berth_lengths
}
\description{
An example of a data frame containing estimated berth lengths.
}
\details{
A data frame containing 8 rows and 6 columns
}
\examples{
berth_lengths

}
\keyword{datasets}
