% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable-specification.R
\name{calculate_journey_specifications}
\alias{calculate_journey_specifications}
\title{Calculate journey specifications}
\usage{
calculate_journey_specifications(timetable, stations)
}
\arguments{
\item{timetable}{Timetable}

\item{stations}{Stations}
}
\description{
Calculate journey specifications
}
