% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-tsars.R
\name{calculate_tsars}
\alias{calculate_tsars}
\title{Calculate TSAR and Sub-Components.}
\usage{
calculate_tsars(track_events, red_events, asset_map)
}
\description{
Calculate berth-level information including TSAR and all sub-components, e.g.,
T_onset, T_clear, T_offset, etc.
}
