% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-data.R
\docType{data}
\name{dwell_times}
\alias{dwell_times}
\title{Example dwell times}
\format{
\describe{
\item{signal}{Signal ID}
\item{berth}{Berth ID}
\item{station}{Station name}
\item{train_id}{Train ID}
\item{group}{Group name}
\item{t_enters}{Date and time that train enters berth}
\item{T_travel}{Berth travel time}
\item{v_entry}{Velocity at entry to berth in m/s}
\item{a_brake}{Braking capacity in m/s^2}
\item{L1}{Distance from start of berth to station}
\item{v_exit}{Velocity at exit of berth in m/s}
\item{a_tract}{Acceleration in m/s^2}
\item{L2}{Distance from station to end of berth}
\item{T_brake}{Time spent braking}
\item{S_brake}{Distance spent braking}
\item{T_const_one}{Time spent moving at a constant speed before
braking for station}
\item{T_move_one}{Total time moving before reaching station}
\item{T_tract}{Time spent accelerating}
\item{S_tract}{Distance spent accelerating}
\item{T_const_two}{Time spent moving at a constant speed before
vacating berth}
\item{T_move_two}{Total time moving after leaving station}
\item{T_dwell}{Dwell time at station}
}
}
\usage{
dwell_times
}
\description{
An example of a data frame containing estimated dwell times
}
\details{
A data frame containing 1,581 rows and 21 columns
}
\examples{
dwell_times

}
\keyword{datasets}
