% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwell-berth_lengths.R
\name{estimate_berth_lengths}
\alias{estimate_berth_lengths}
\title{Estimate berth lengths}
\usage{
estimate_berth_lengths(
  timetable_specification,
  berth_events_groups,
  id_matching,
  distance.miles,
  speed.miles
)
}
\arguments{
\item{timetable_specification}{Timetable specification}

\item{berth_events_groups}{Berth events groups}

\item{id_matching}{Id matching}

\item{distance.miles}{Distance in miles}

\item{speed.miles}{Speed in miles}
}
\description{
Estimate berth lengths
}
