% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{filter_relevant_services}
\alias{filter_relevant_services}
\title{Filter timetable by stations}
\usage{
filter_relevant_services(timetable, stations)
}
\arguments{
\item{timetable}{Data frame containing timetable data.}

\item{stations}{List containing station names.}
}
\value{
A (usually) smaller timetable only containing trains which pass
through or stop at the specified stations.
}
\description{
Filter timetable by stations
}
