% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{filter_relevant_services}
\alias{filter_relevant_services}
\alias{filter_relevant_direction}
\title{Filter timetable to relevant trains}
\usage{
filter_relevant_services(timetable, stations)

filter_relevant_direction(timetable, stations)
}
\arguments{
\item{timetable}{Data frame containing timetable data. Strict input validation
is performed. Must be a data frame with the following columns:
\itemize{
\item{\code{train_header}}{\code{\link[=character]{character()}} Train identifier}
\item{\code{dt_origin}}{\code{\link[lubridate:posix_utils]{lubridate::POSIXct()}} Datetime at which the train
originated}
\item{\code{geo}}{\code{\link[=character]{character()}} Name of event location}
\item{\code{event}}{\code{\link[=character]{character()}} Type of event, must only contain the
following values: \code{Pass}, \code{Arrive}, \code{Depart}, \code{Originate},
\code{Terminate}}
\item{\code{wtt}}{\code{\link[lubridate:posix_utils]{lubridate::POSIXct()}} Scheduled datetime of event}
\item{\code{t}}{\code{\link[lubridate:posix_utils]{lubridate::POSIXct()}} Actual datetime of event}
\item{\code{delay}}{\code{\link[=numeric]{numeric()}} Difference between actual time and scheduled
time of event}
\item{\code{allow}}{\code{\link[=numeric]{numeric()}} Timetabled delay allowance}
}}

\item{stations}{A list of TIPLOCs (\code{geo}) through which trains pass. This list
is used to work out which trains travel through the track section of
interest. For example, there could be a fork in the track, where trains can
travel to the left or to the right. If you want to only look at trains which
turn to the left, \code{stations} would include all the TIPLOCs of interest which
are situated along the left track. Any trains which do not pass through any
of the TIPLOCs specified in \code{stations} will be discarded.}
}
\value{
\code{filter_relevant_services} returns a (usually) smaller timetable
only containing trains which pass through or stop at the specified
stations.

\code{filter_relevant_direction} returns a (usually) smaller timetable only
containing trains which pass through or stop at the specified stations in
the given order.
}
\description{
\code{filter_relevant_services()} filters only by station, ignoring
direction.

\code{filter_relevant_direction()} filters by both station and direction. It is
assumed that \code{stations} is in the forward direction, i.e., the first
element is at the start of the track section and the last element is at the
end.
}
\examples{
data(timetable, stations)
timetable
timetable_services <- filter_relevant_services(timetable, stations)
timetable_services
timetable_direction <- filter_relevant_direction(timetable, stations)
timetable_direction

}
\seealso{
\code{\link[=wrangle_timetable]{wrangle_timetable()}} \code{\link[=find_calling_patterns]{find_calling_patterns()}}
}
