% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{find_calling_patterns}
\alias{find_calling_patterns}
\title{Find calling patterns from timetable}
\usage{
find_calling_patterns(timetable_raw, stopping_stations)
}
\arguments{
\item{stopping_stations}{a character vector containing the names of stations
at which trains stop. Any trains which do not stop at any of these stations
will be labeled as 'fast' (non-stopping) trains. Trains which do stop at
these stations will either be labeled as 'stopping-all' or they will be
labeled with the specific stations at which they stop.}
}
\value{
The timetable with calling patterns added.
}
\description{
Find calling patterns from timetable
}
\examples{
data(timetable, stations, stopping_stations)
calling_patterns <- timetable |>
                      filter_relevant_direction(stations) |>
                      find_calling_patterns(stopping_stations)
calling_patterns[,c("train_header", "dt_origin", "geo", "event", "group")]

}
\seealso{
\code{\link[=filter_relevant_services]{filter_relevant_services()}} \code{\link[=wrangle_timetable]{wrangle_timetable()}}
}
