% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R, R/wrangle_timetable.R
\name{get_event_mapping}
\alias{get_event_mapping}
\title{Get Timetable Event Mapping from Environment}
\usage{
get_event_mapping()

get_event_mapping()
}
\value{
A data frame containing a 1-1 mapping of timetable event codes to
readable names. These are as follows:
\itemize{
\item O = Originate
\item P = Pass
\item A = Arrive
\item D = Depart
\item T = Terminate
}

The event mapping table as a tibble data frame.
}
\description{
This function retrieves the event mapping from the current environment. The
event mapping defines the possible timetable events.

This function retrieves the current event mapping table.
}
\details{
The event mapping table is stored in the environment object named "env".
}
\seealso{
\link{set_event_mapping}
}
