% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_timetable.R
\name{get_stations}
\alias{get_stations}
\title{Get Stations}
\usage{
get_stations()
}
\value{
A character vector containing the stations.
}
\description{
This function retrieves the stations from the environment.
}
\examples{
# Create an environment containing the stations
env <- new.env()
env$stations <- c("Station A", "Station B", "Station C")

# Call the function to get the stations
get_stations()

}
\seealso{
\link{set_stations}
}
