% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-clustering.R
\name{plot_clusters}
\alias{plot_clusters}
\title{Plot clustered train journeys}
\usage{
plot_clusters(cluster_events)
}
\arguments{
\item{cluster_events}{Data frame matching output from \code{\link[=cluster_journeys]{cluster_journeys()}}}
}
\description{
Plot clustered train journeys
}
