% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_centrix.R
\name{preprocess_signal_events}
\alias{preprocess_signal_events}
\title{Preprocess raw signal events}
\usage{
preprocess_signal_events(raw_signal_events, state_mapping)
}
\arguments{
\item{raw_signal_events}{Raw signal data.}

\item{state_mapping}{Data frame defining how to convert signal states to
aspects.}
}
\description{
Preprocess raw signal events
}
