% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_centrix.R
\name{preprocess_signal_events}
\alias{preprocess_signal_events}
\title{Preprocess raw signal events}
\usage{
preprocess_signal_events(raw_signal_events)
}
\arguments{
\item{raw_signal_events}{Raw signal data.}
}
\description{
Preprocess raw signal events
}
