% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_matching.R
\name{preprocess_timetable}
\alias{preprocess_timetable}
\title{Preprocess Timetable Data for ID Matching}
\usage{
preprocess_timetable(timetable, train_classes)
}
\arguments{
\item{timetable}{Data frame containing timetable data}

\item{train_classes}{Data frame containing train classes}
}
\description{
Preprocess Timetable Data for ID Matching
}
