% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_timetable.R
\name{preprocess_timetable}
\alias{preprocess_timetable}
\title{Preprocess Timetable}
\usage{
preprocess_timetable(path, names, ...)
}
\arguments{
\item{path}{The path to the directory containing the Excel files.}

\item{names}{A named character vector representing column name mappings for
the timetable. The names should correspond to the column names in the Excel
files. The names of the vector will be used as the target column names, and
the values as the corresponding column names in the Excel files. The
columns not listed in \code{names} will be ignored in the final timetable.}

\item{...}{Additional arguments passed to the \link{read_excel_files} function.}
}
\value{
A preprocessed timetable with selected and renamed columns.
}
\description{
This function reads timetable data from Excel files, performs data
preprocessing, and returns a processed timetable.
}
\details{
The function reads Excel files from the specified directory using
\link{read_excel_files}. It then performs the following preprocessing steps:
\itemize{
\item{1. Selects columns based on the names provided in the \code{names}
parameter. Columns with names that match the provided names (partially or
fully) will be selected.}
\item{2. Renames the selected columns using the corresponding names provided
in \code{names}.} \item{3. Selects specific columns from the processed data to
include in the final timetable.} \item{4. Performs a left join with the
event mapping table retrieved using \link{get_event_mapping}. The event mapping
table is stored in the environment object named "env". The event mapping
should have two columns: "event" and "name".} \item{5. Renames the "name"
column from the event mapping table to "event" in the timetable.} }
}
\seealso{
\link{read_excel_files}, \link{get_event_mapping}
}
