% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-preprocessing.R
\name{preprocess_track_events}
\alias{preprocess_track_events}
\title{Preprocess Raw Track Events}
\usage{
preprocess_track_events(raw_track_events, asset_map)
}
\arguments{
\item{raw_track_events}{Data frame containing raw track events from Centrix
data. No input validation is performed. The data should matche the
structure of Centrix data as described in \code{\link[=wrangle_centrix]{wrangle_centrix()}}.}

\item{asset_map}{Data frame containing a map of the track section, used to
filter the raw track events down to only the tracks specified in the asset
map. No input validation is performed, see \code{\link[=wrangle_centrix]{wrangle_centrix()}} for the
expected structure.}
}
\value{
A data frame containing track events with columns:
\itemize{
\item 'track': (character) track ID,
\item 'dt': (lubridate::POSIXct) datetime,
\item 'occupied': (logical) TRUE if train enters track, else FALSE,
\item 'event': (character) 'enters' if train enters track, else 'vacates'.
}
}
\description{
Performs preprocessing on raw track events. Converts transition data to
whether the train enters or vacates the track.
}
\seealso{
\code{\link[=wrangle_centrix]{wrangle_centrix()}}
}
