% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_centrix.R
\name{preprocess_track_events}
\alias{preprocess_track_events}
\title{Preprocess raw track data}
\usage{
preprocess_track_events(raw_track_events)
}
\arguments{
\item{raw_track_events}{Data frame containing raw track data}
}
\description{
Preprocess raw track data
}
