% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_centrix}
\alias{raw_centrix}
\title{Example Centrix data}
\format{
\describe{
\item{asset}{Centrix asset ID, signal assets include state data.}
\item{dt}{Date and time of event.}
\item{transition}{Event transition data.}
}
}
\usage{
raw_centrix
}
\description{
An example of a data frame containing raw Centrix data. This is real data
containing 15 days of Centrix data. The data have been anonymised.
}
\details{
A data frame containing 110,339 rows and 3 columns.
}
\examples{
raw_centrix

}
\keyword{datasets}
