% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_csv_files}
\alias{read_csv_files}
\title{Read raw CSV data from multiple files}
\usage{
read_csv_files(path, ...)
}
\arguments{
\item{path}{A character string specifying the path to the directory
containing the CSV files}

\item{...}{Additional arguments to be passed to \link[readr]{read_csv}}
}
\description{
Reads all the CSV files contained in the directory specified  by \code{path} and
combines them into a single data frame.
}
\examples{
# Create a temporary directory for the example
tempdir <- tempdir()
tempdir <- file.path(tempdir, "example")
dir.create(tempdir)

# Generate a sample data frame
df <- data.frame(
x = c("a", "b", "c"),
y = c(1, 2, 3)
)

# Define the file path
file_path <- file.path(tempdir, "example.csv")

# Save the data frame as a CSV file
write.csv(df, file = file_path, row.names = FALSE)

# Define column names and types
names <- c("x", "y")
types <- readr::cols(
"x" = readr::col_character(),
"y" = readr::col_integer()
)

# Read the data using read_centrix()
read_csv_files(tempdir, names, types)

# Delete the temporary directory
unlink(tempdir, recursive = TRUE)

}
