% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_files.R
\name{read_excel_files}
\alias{read_excel_files}
\title{Read raw excel data from multiple files}
\usage{
read_excel_files(path, progress = TRUE, ...)
}
\arguments{
\item{path}{A string pointing to the directory containing excel files}

\item{progress}{Whether to show a progress bar, see \code{\link{map}}}

\item{...}{Additional arguments to be passed to \code{\link{read_excel}}}
}
\description{
Read raw excel data from multiple files
}
