% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_timetable.R
\name{set_stations}
\alias{set_stations}
\title{Set Stations}
\usage{
set_stations(stations)
}
\arguments{
\item{stations}{A character vector containing the names of the stations.}
}
\description{
This function sets the stations in the environment.
}
\details{
The function validates the input stations to ensure they meet the
requirements: 1. The "stations" argument must be a non-empty character
vector. 2. Each element of the character vector must be a string (i.e., of
character type). If the input stations do not meet these conditions, the
function will raise an error.
}
\examples{
# Set stations using the function
set_stations(list('Station A', 'Station B', 'Station C'))

}
\seealso{
\link{get_stations}
}
