% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix-preprocessing.R
\name{split_signal_track_events}
\alias{split_signal_track_events}
\title{Split Signal and Track Events}
\usage{
split_signal_track_events(raw_events)
}
\arguments{
\item{raw_events}{A data frame containing raw Centrix data with both signal
and track events. No input validation is performed, see \code{\link[=wrangle_centrix]{wrangle_centrix()}}
for expected structure.}
}
\value{
A list containing two data frames. The first data frame contains raw
signal events, the second contains raw track events.
}
\description{
Splits a single Centrix data frame containing both signal and track events
into a list containing two data frames.
}
\seealso{
\code{\link[=wrangle_centrix]{wrangle_centrix()}}
}
