% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable-data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Example stations}
\format{
An object of class \code{list} of length 6.
}
\usage{
stations
}
\description{
List of example stations. This is used for timetable wrangling
}
\details{
A list with 6 elements.
}
\examples{
stations

}
\seealso{
\code{\link[=wrangle_timetable]{wrangle_timetable()}}
}
\keyword{datasets}
