% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{time_windows}
\alias{time_windows}
\title{Example time windows}
\format{
\describe{
\item{window}{Window ID}
\item{interval}{Window time interval}
}
}
\usage{
time_windows
}
\description{
An example of time windows calculated from Centrix data denoting time
intervals of valid train journeys.
}
\details{
A data frame containing 681 rows and 2 columns
}
\examples{
time_windows

}
\keyword{datasets}
