% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_centrix.R
\name{wrangle_centrix}
\alias{wrangle_centrix}
\title{Wrangle raw Centrix data}
\usage{
wrangle_centrix(signal_events, track_events)
}
\arguments{
\item{signal_events}{Data frame containing pre-processed signal data.}

\item{track_events}{Data frame containing pre-processed track data.}
}
\description{
Wrangle raw Centrix data
}
