% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrix.R
\name{wrangle_centrix}
\alias{wrangle_centrix}
\title{Wrangle Raw Centrix Data}
\usage{
wrangle_centrix(raw_centrix, asset_map, state_map = state_mapping)
}
\arguments{
\item{raw_centrix}{A data frame containing raw Centrix data. Strict input
validation is applied to ensure the data adhere to the expected
structure. The data should be in a data frame with the following columns:
\itemize{
\item{\code{asset}}: A character vector containing the asset ID. Tracks
are expected to follow the regex pattern \code{"T[A-Z]+(-[0-9])?"}. E.g.,
'TABC' 'TABC-1' 'TABC-2'.
Signals are expected to follow the regex pattern \code{"S[0-9]+\\s[A-Z]+"}.
E.g., 'S123 HGE' 'S123 I'.
The second part of the signal ID should be a state code referring to
the signal aspect, e.g. 'RGE' for red or 'DGE' for green. For other
codes see:
\url{https://wiki.openraildata.com/index.php/Signalling_Nomenclature#Signals}.
\item{\code{dt}}: A \link[lubridate:posix_utils]{lubridate::POSIXct} object representing the date
and time at which the observation was made.
\item{\code{transition}}: A character vector representing the state
transition that caused the observation. A transition can be one of
"UP to DN" or "DN to UP", anything else will be ignored. For tracks,
"UP to DN" signifies a train entering the track, while "DN to UP" is a
train vacating the track. For signals, "UP to DN" represents an aspect
change. Therefore, any signals with a "DN to UP" transition will be
ignored.
}}

\item{asset_map}{A data frame containing a 1-1 mapping from signals to berths
and a 1-many mapping from berths to tracks. The map also contains a 1-many
mapping from tracks to events, reflecting the fact that there is an 'enter'
and a 'vacate' event for each track. Therefore, the data frame should
adhere to the following structure:
\itemize{
\item{\code{signal}}: A character vector containing the signal ID. It
should follow the regex pattern \code{"S[0-9]+"}.
\item{\code{berth}}: A character vector containing the berth ID. It
should follow the regex pattern \code{"[A-Z]+"}.
\item{\code{track}}: A character vector containing the track ID. It
should follow the regex pattern \code{"T[A-Z]+(-[0-9])?"}.
\item{\code{event}}: A character vector containing the event type. It
must be one of either "enters" or "vacates".
}}

\item{state_map}{A data frame containing a 1-1 mapping from signal state
to signal aspect. See \link{state_mapping} (the default) for the expected
structure.}
}
\value{
A data frame containing the fully processed Centrix data, containing
berth-level information about signal and track events. This includes TSARs
and all sub-components, e.g. 'T_travel' and 'T_offset'.
}
\description{
This function takes in raw Centrix data as an input and returns a fully
processed data frame containing berth-level signal and track observations,
along with calculations of TSAR and all its subcomponents.
}
\examples{
# This will require some example data

}
\seealso{
All lower-level functions that this function wraps.
}
