% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{wrangle_timetable}
\alias{wrangle_timetable}
\title{Wrangle Timetable Data}
\usage{
wrangle_timetable(timetable, stations)
}
\arguments{
\item{timetable}{Data frame containing timetable data.}

\item{stations}{List containing station names.}
}
\description{
Wrangle Timetable Data
}
