% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{wrangle_timetable}
\alias{wrangle_timetable}
\title{Wrangle Timetable Data}
\usage{
wrangle_timetable(timetable, stations)
}
\arguments{
\item{timetable}{Data frame containing timetable data.}

\item{stations}{List containing station names.}
}
\value{
A subset of the timetable which contains trains that pass through the
specified stations in the given order. Time zone is set to "UTC".
}
\description{
A wrapper for timetable processing functions \code{\link[=filter_relevant_services]{filter_relevant_services()}},
\code{\link[=filter_relevant_direction]{filter_relevant_direction()}}, and \code{\link[=find_calling_patterns]{find_calling_patterns()}}.
}
\seealso{
\code{\link[=filter_relevant_services]{filter_relevant_services()}} \code{\link[=filter_relevant_direction]{filter_relevant_direction()}}
\code{\link[=find_calling_patterns]{find_calling_patterns()}}
}
