% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{wrangle_timetable}
\alias{wrangle_timetable}
\title{Wrangle Timetable Data}
\usage{
wrangle_timetable(timetable, stations, stopping_stations)
}
\arguments{
\item{timetable}{Data frame containing timetable data. Strict input validation
is performed. Must be a data frame with the following columns:
\itemize{
\item{\code{train_header}}{\code{\link[=character]{character()}} Train identifier}
\item{\code{dt_origin}}{\code{\link[lubridate:posix_utils]{lubridate::POSIXct()}} Datetime at which the train
originated}
\item{\code{geo}}{\code{\link[=character]{character()}} Name of event location}
\item{\code{event}}{\code{\link[=character]{character()}} Type of event, must only contain the
following values: \code{Pass}, \code{Arrive}, \code{Depart}, \code{Originate},
\code{Terminate}}
\item{\code{wtt}}{\code{\link[lubridate:posix_utils]{lubridate::POSIXct()}} Scheduled datetime of event}
\item{\code{t}}{\code{\link[lubridate:posix_utils]{lubridate::POSIXct()}} Actual datetime of event}
\item{\code{delay}}{\code{\link[=numeric]{numeric()}} Difference between actual time and scheduled
time of event}
\item{\code{allow}}{\code{\link[=numeric]{numeric()}} Timetabled delay allowance}
}}

\item{stations}{A list of TIPLOCs (\code{geo}) through which trains pass. This list
is used to work out which trains travel through the track section of
interest. For example, there could be a fork in the track, where trains can
travel to the left or to the right. If you want to only look at trains which
turn to the left, \code{stations} would include all the TIPLOCs of interest which
are situated along the left track. Any trains which do not pass through any
of the TIPLOCs specified in \code{stations} will be discarded.}

\item{stopping_stations}{A subset of \code{stations}, which specifies which TIPLOCs
trains can stop at. This is used for calculating calling patterns.
Therefore, any train which does not stop at any of the stations specified in
\code{stopping_stations} will be labeled as a 'fast' train. Trains which do stop
at these stations will be labeled with 'stopping-', with the specific
stations at which they stop attached as suffixes. For example, if a train
stops at 'geo10', its calling pattern will be labeled as 'stopping-geo10'.}
}
\value{
A subset of the timetable which contains trains that pass through the
specified stations in the given order. Time zone is set to "UTC".
}
\description{
A wrapper for timetable processing functions \code{\link[=filter_relevant_services]{filter_relevant_services()}},
\code{\link[=filter_relevant_direction]{filter_relevant_direction()}}, and \code{\link[=find_calling_patterns]{find_calling_patterns()}}.
}
\examples{
data(timetable, stations, stopping_stations)
timetable
timetable_subset <- wrangle_timetable(timetable, stations, stopping_stations)
timetable_subset

}
\seealso{
\code{\link[=filter_relevant_services]{filter_relevant_services()}} \code{\link[=find_calling_patterns]{find_calling_patterns()}}
}
