% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_timetable.R
\name{wrangle_timetable}
\alias{wrangle_timetable}
\title{Identifies the services that travel through the stations from start to end}
\usage{
wrangle_timetable(timetable)
}
\arguments{
\item{timetable}{Pre-processed timetable data}
}
\description{
Identifies the services that travel through the stations from start to end
}
